+++ lib/SILOptimizer/UtilityPasses/SerializeSILPass.cpp
@@ -451,21 +451,9 @@ public:
         !M.getOptions().CrossModuleOptimization)
       return;
 
-    // Mark all reachable functions as "anchors" so that they are not
-    // removed later by the dead function elimination pass. This
-    // is required, because clients may reference any of the
-    // serialized functions or anything referenced from them. Therefore,
-    // to avoid linker errors, the object file of the current module should
-    // contain all the symbols which were alive at the time of serialization.
     LLVM_DEBUG(llvm::dbgs() << "Serializing SILModule in SerializeSILPass\n");
     M.serialize();
 
-    // If we are not optimizing, do not strip the [serialized] flag. We *could*
-    // do this since after serializing [serialized] is irrelevent. But this
-    // would incur an unnecessary compile time cost since if we are not
-    // optimizing we are not going to perform any sort of DFE.
-    if (!getOptions().shouldOptimize())
-      return;
     removeSerializedFlagFromAllFunctions(M);
   }
 };
