+++ test/SIL/Serialization/public_non_abi.sil
@@ -20,11 +20,11 @@ bb0:
 }
 
 // Make sure the function body is deserialized.
-// CHECK-LABEL: sil shared_external [serialized] @public_non_abi_function : $@convention(thin) () -> ()
+// CHECK-LABEL: sil shared_external @public_non_abi_function : $@convention(thin) () -> ()
 // CHECK: function_ref @other_public_non_abi_function
 // CHECK: return
 sil hidden_external [serialized] @public_non_abi_function : $@convention(thin) () -> ()
 
 // Make sure the function body is deserialized.
-// CHECK-LABEL: sil shared_external [serialized] @other_public_non_abi_function : $@convention(thin) () -> ()
+// CHECK-LABEL: sil shared_external @other_public_non_abi_function : $@convention(thin) () -> ()
 // CHECK: return
