+++ test/SILOptimizer/Inputs/cast_folding_opaque_other.swift
@@ -1,0 +1,8 @@
+public protocol P {}
+
+public struct Underlying : P {
+}
+
+public func returnOpaque() -> some P {
+  return Underlying()
+}
