+++ test/SILOptimizer/cast_folding_opaque.swift
@@ -1,16 +1,7 @@
-// RUN: %target-swift-frontend -enable-library-evolution -disable-availability-checking -O -emit-sil %s
-// RUN: %target-swift-frontend -enable-library-evolution -disable-availability-checking -Onone -emit-sil %s | %FileCheck %s
+// RUN: %target-swift-frontend -enable-library-evolution -disable-availability-checking -O -emit-sil %s %S/Inputs/cast_folding_opaque_other.swift -module-name cast_folding_opaque
+// RUN: %target-swift-frontend -enable-library-evolution -disable-availability-checking -Onone -emit-sil -primary-file %s %S/Inputs/cast_folding_opaque_other.swift -module-name cast_folding_opaque | %FileCheck %s
 
-public protocol P {}
-
-public struct Underlying : P {
-}
-
-public func returnOpaque() -> some P {
-  return Underlying()
-}
-
-// CHECK-LABEL: sil [serialized] @$s19cast_folding_opaque23testCastOpaqueArchetypeAA10UnderlyingVyF
+// CHECK-LABEL: sil @$s19cast_folding_opaque23testCastOpaqueArchetypeAA10UnderlyingVyF
 // CHECK:   [[O:%.*]] = alloc_stack $@_opaqueReturnTypeOf("$s19cast_folding_opaque12returnOpaqueQryF", 0)
 // CHECK:   [[F:%.*]] = function_ref @$s19cast_folding_opaque12returnOpaqueQryF
 // CHECK:   apply [[F]]([[O]])
