+++ test/SILOptimizer/linker.swift
@@ -3,7 +3,7 @@
 // RUN: %target-swift-frontend %s -I %t -emit-sil | %FileCheck %s
 // RUN: %target-swift-frontend %s -I %t -O -emit-sil | %FileCheck %s --check-prefix=OPT
 
-// CHECK: sil [serialized] [noinline] @$ss11doSomethingyyF : $@convention(thin) () -> (){{$}}
+// CHECK: sil [noinline] @$ss11doSomethingyyF : $@convention(thin) () -> (){{$}}
 // OPT: sil public_external [noinline] @$ss11doSomethingyyF : $@convention(thin) () -> () {
 doSomething()
 
@@ -11,7 +11,7 @@ doSomething()
 // CHECK-NOT: return
 doSomething2()
 
-// CHECK: sil [serialized] [noinline] @$ss16callDoSomething3yyF : $@convention(thin) () -> (){{$}}
+// CHECK: sil [noinline] @$ss16callDoSomething3yyF : $@convention(thin) () -> (){{$}}
 // OPT: sil public_external [noinline] @$ss16callDoSomething3yyF : $@convention(thin) () -> () {
 
 // OPT: sil @unknown
